separating common abstract-element properties
